services/authorizer/pyproject.toml
[tool.poetry]
name = "mavik-authorizer"
version = "0.1.0"
description = "AWS Lambda Authorizer for Mavik AI API Gateway"
authors = ["Mavik Capital <engineering@mavikcapital.com>"]
packages = [
    { include = "authorizer", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^41.0.7"
requests = "^2.31.0"
pydantic = "^2.5.0"

# Local packages
mavik-common = {path = "../../packages/common", develop = true}
mavik-config = {path = "../../packages/config", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
black = "^23.11.0"
mypy = "^1.7.1"
ruff = "^0.1.6"

[tool.poetry.group.test.dependencies]
moto = "^4.2.14"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=authorizer",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80"
]